# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.DeviceType:
    description: DeviceType is a struct for device type
    properties:
      id:
        description: ID is an int32 for device type id
        example: 1
        type: integer
      name:
        description: Name is a string for device type name
        example: computer
        type: string
      status:
        description: Status is a string for device type status either active or inactive
        example: active|inactive
        type: string
    type: object
...
  models.JsonResponse:
    description: JsonResponse is a struct for response JSON message
    properties:
      Action:
        description: Action is a string for response action
        type: string
      MSG:
        description: Message is an interface for response message can be string, models.DeviceType,
          models.Manufacturer, models.Equipment
      Status:
        description: Status is a string for response status
        example: SUCCESS|ERROR
        type: string
    type: object
...
info:
  contact: {}
  description: This is the API to interact with Equipment database
  termsOfService: http://swagger.io/terms/
  title: Equipment API
  version: "1.0"
paths:
  /device:
    get:
      consumes:
      - application/json
      description: get all device types from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.JsonResponse'
            - properties:
                MSG:
                  $ref: '#/definitions/models.DeviceType'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.JsonResponse'
      summary: get all device types
      tags:
      - device
      x-order: 1
    post:
      consumes:
      - application/json
      description: create device type for the database
      parameters:
      - description: Device Name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.JsonResponse'
      summary: create device type
      tags:
      - device
      x-order: 4
  /device/{id}:
    get:
      consumes:
      - application/json
      description: get device type by ID from the database
      parameters:
      - description: Device ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.JsonResponse'
            - properties:
                MSG:
                  $ref: '#/definitions/models.DeviceType'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.JsonResponse'
      summary: get device type by ID
      tags:
      - device
      x-order: 2
  /device/{id}/name:
    patch:
      consumes:
      - application/json
      description: update device type by name ID from the database
      parameters:
      - description: Device ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Device Name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.JsonResponse'
      summary: update device type by name ID
      tags:
      - device
      x-order: 3
  /device/{id}/status:
    patch:
      consumes:
      - application/json
      description: update device type by status ID from the database
      parameters:
      - description: Device ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Device Status
        enum:
        - active
        - inactive
        in: query
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.JsonResponse'
      summary: update device type by status ID
      tags:
      - device
      x-order: 3
...
swagger: "2.0"